+++ examples/histogram-parallel/main.go
@@ -3,93 +3,76 @@ package main
 import (
 	// import the entire framework (including bundled verilog)
 	_ "sdaccel"
+	"sdaccel/memory"
 )
-
-const (
-	INPUT_INDEX = iota
-	OUTPUT_INDEX
-	LENGTH_INDEX
-)
-
-type Histogram [512]uint32
-
-func (hist Histogram) Add(sample uint32) {
-	hist[sample>>(32-9)] += 1
-}
-
-func (hist Histogram) ToBytes() [4][512]byte {
-	var ret [4][512]byte
-	for i := 0; i < 4; i++ {
-		for j := 0; j < 512; j += 4 {
-			k := j / 4
-			val := hist[i<<7|k]
-			ret[i][k] = byte(val >> 24)
-			ret[i][k+1] = byte(val >> 16)
-			ret[i][k+2] = byte(val >> 8)
-			ret[i][k+3] = byte(val)
-		}
-	}
-	return ret
-}
 
 // magic identifier for exporting
 func Top(
-	controlAddr chan<- uint32,
-	controlData <-chan uint32) {
+	inputData uintptr,
+	outputData uintptr,
+	length uint32,
 
-	readParam := func(a uint32) uint32 {
-		controlAddr <- a
-		return <-controlData
-	}
+	memReadAddr chan<- memory.Addr,
+	memReadData <-chan memory.ReadData,
 
-	readAddressParam := func(a uint32) uint64 {
-		controlAddr <- a
-		top := <-controlData
-		controlAddr <- a + 1
-		return (uint64(top) << 32) + uint64(<-controlData)
-	}
+	memWriteAddr chan<- memory.Addr,
+	memWriteData chan<- memory.WriteData,
+	memResp <-chan memory.Response) {
 
-	readMemory := func(a uint64) [512]byte {
-		return [512]byte{}
-	}
+	readChan := make(chan uintptr)
+	readRespChan := make(chan uint32)
 
-	writeMemory := func(a uint64, bytes [512]byte) {
-	}
+	incrChan := make(chan uintptr)
+	incrResp := make(chan uint32)
 
-	inputData := readAddressParam(INPUT_INDEX)
-	outputData := readAddressParam(INPUT_INDEX)
+	go func() {
+		for {
+			var addr uintptr
 
-	// how many samples are there? (uint32)
-	length := readParam(LENGTH_INDEX)
+			// This block provides single access to memory, with two
+			// separate operations: reading or incrementing a pointer
+			// in memory.
 
-	var histogram Histogram
-
-	for ; length > 0; length -= 16 {
-		toProcess := length
-
-		if toProcess > 16 {
-			toProcess = 16
+			// All access is through channels allowing operations to
+			// be enqueued, only blocking when the response is needed
+			select {
+			case addr = <-readChan:
+				readRespChan <- memory.Read(addr, memReadAddr, memReadData)
+			case addr = <-incrChan:
+				current := memory.Read(addr, memReadAddr, memReadData)
+				memory.Write(addr, current+1, memWriteAddr, memWriteData, memResp)
+				incrResp <- current + 1
+			}
 		}
+	}()
 
-		bytes := readMemory(inputData)
-
-		inputData += 1
-
-		for i := uint32(0); i < toProcess; i++ {
-			sample :=
-				uint32(bytes[i])<<24 |
-					uint32(bytes[i+1])<<16 |
-					uint32(bytes[i+2])<<8 |
-					uint32(bytes[i+3])
-
-			histogram.Add(sample)
+	go func() {
+		// length is the number of addresses we are supposed to read
+		// so this block enqueues each address to read.
+		for i := length; i > 0; i-- {
+			readChan <- inputData
+			inputData += 4
 		}
+	}()
 
+	go func() {
+		for i := length; i > 0; i-- {
+			// get the read response that was previously enqueued.
+			sample := <-readRespChan
+			// If we think of external memory we are writing to as a [512]uint32, this would be the index we access
+			index := uint16(sample) >> (16 - 9)
+			pointerDiff := index << 2
+			// And this is that index as a pointer to external memory
+			outputPointer := outputData + uintptr(pointerDiff)
+			// enqueue an increment operation on that pointer
+			incrChan <- outputPointer
+		}
+	}()
+
+	// Wait for each response for increment operations
+	for i := length; i > 0; i-- {
+		<-incrResp
 	}
 
-	bytes := histogram.ToBytes()
-
-	for i := 0; i <= 4; i++ {
-		writeMemory(outputData+uint64(i), bytes[i])
-	}
+	// Once that's done, we can exit.
 }